Added attribute for download url to use
Created default attribute [:distelli][:agent][:download_url] to be used for
setting the download url for the agent. This way on-prem distelli instances can
use their instance to download the agent. This has the affect of not being
dependend on distelli.com, and also ensuring that the latest agent release is
always compatible with the on-prem distelli version.